+++ DatadogCore/Tests/Objc/ObjcAPITests/DDConfiguration+apiTests.m
@@ -69,6 +69,8 @@
     configuration.additionalConfiguration = @{@"additional": @"config"};
     [configuration setEncryption:[CustomDDDataEncryption new]];
     configuration.backgroundTasksEnabled = true;
+    [configuration setRemoteConfiguration:[[DDRemoteConfiguration alloc] initWithId:@"abc-123"]];
+    [configuration setRemoteConfiguration:nil];
 }
 
 - (void)testDatadogCrashReporterAPI {